Define simplicity_{alloc,free}
These macros can be overwritten to point to a Rust function that does
the (de)allocation. The Rust (de)allocator needs to know the memory
alignment of the (de)allocated type, so this is an parameter of the
macro. It is the responsibility of the caller to ensure correct
alignment.

The default C alloc ignores the alignment parameter, so alignment errors
might not be caught. We could use aligned_alloc from C11 in the C code.
IDK how we could statically assert correct alignment in debug builds,
but if we can then we should.

The default C free ignores both the size and the alignment parameter,
which is even more dangerous.