Return number of allocated bytes of elements env
Rust needs to know the number of allocated bytes to free the Elements
transaction and the Taproot environment. I generalized the allocation
functions to include a buffer that is filled with the number of
allocated bytes. Including such a buffer seems to be uncommon in C, so
the generalized function is inlined and effectively invisible to C land.
Rust can define a custom C function and call the inlined function to
make use of it.